Conflicts:
	app/controllers/incident_reports_controller.rb

+++ app/controllers/incident_reports_controller.rb
@@ -1,8 +1,8 @@
 class IncidentReportsController < ApplicationController
-  before_filter :admin_authorize, :except => [:new_report, :show, :edit]
-  before_filter :general_authorize  
+  before_filter :admin_assistant_authorize_view_access, :except => [:new_report, :show, :edit, :create, :update_participant_list, :update, :destroy]
+  before_filter :not_admin_assistant_authorize_view_access, :except => [:show, :index]
   skip_before_filter :verify_authenticity_token
-  acts_as_iphone_controller = true
+
   
   
   # GET /incident_reports
@@ -32,12 +32,20 @@ class IncidentReportsController < ApplicationController
     # get the report for the view to show
     @incident_report = IncidentReport.find(params[:id])
     
+    if (@incident_report.submitted? && @incident_report.updated_at + 1.minutes < Time.now && current_staff.access_level == Authorize.ra_access_level) || (@incident_report.staff != current_staff && current_staff.access_level == Authorize.ra_access_level)
+        flash[:notice] = "Unauthorized Access"
+        redirect_to "/home/landingpage"
+        return
+    end
+      
+    
     self.clear_session #probably not necessary, but good practice anyway
     
     respond_to do |format|
       format.html # show.html.erb
       format.xml  { render :xml => @incident_report }
       #format.iphone {render :layout => false}
+      format.iphone {render :layout => 'mobile_application'}
     end
   end
   
@@ -48,23 +56,33 @@ class IncidentReportsController < ApplicationController
   
   # GET /incident_reports/1/edit
   def edit
-    if(session[:incident_report] !=nil)
-      @incident_report = session[:incident_report]
-      @annotation = session[:annotation]
-    else
+
       # get the report and annotation for the view to edit
       @incident_report = IncidentReport.find(params[:id])
+      
+      if (@incident_report.submitted? && current_staff.access_level == Authorize.ra_access_level) || (!@incident_report.submitted? && current_staff.access_level == Authorize.ra_access_level && @incident_report.staff != current_staff)
+        flash[:notice] = "Unauthorized Access"
+        redirect_to "/home/landingpage"
+        return
+      end
+      
       @annotation = Annotation.find(@incident_report.annotation_id)
       
       # save the report and annotation into the session
       session[:incident_report] = @incident_report
       session[:annotation] = @annotation
-    end
+
     
     # if this is the return from a report_search page
     if session[:students] != nil
       @incident_report.add_default_report_student_relationships_for_participant_array(session[:students])
     end 
+    
+    respond_to do |format|
+        format.html 
+        format.iphone {render :layout => 'mobile_application'}
+      end
+      
   end
   
   
@@ -75,7 +93,29 @@ class IncidentReportsController < ApplicationController
   # POST /incident_reports
   # POST /incident_reports.xml
   def create
-
+    #search_submit is name of button clicked called "add students"
+    if params[:search_submit] != nil
+      # update values of incident_report
+      @incident_report = session[:incident_report]
+      @incident_report.update_attributes_without_saving(params[:incident_report])
+      
+      # process parameters into reported infractions
+      self.add_reported_infractions_to_report(@incident_report, params)  
+      
+      
+      # update annotation
+      @annotation = session[:annotation]
+      @annotation.text = params[:annotation]
+      
+      # render search page, sending own path as return address
+      respond_to do |format|
+        format.html { redirect_to '/search/report_search?/incident_reports/new_report' }
+        format.xml  { render :xml => @incident_report, :status => :created, :location => @incident_report }
+        format.iphone {redirect_to '/search/report_search.iphone?/incident_reports/new_report.iphone', :layout => 'mobile_application'}
+      end
+      
+      
+    else
       # any submit button except search_submit (save_submit or submit_submit)
       @incident_report = session[:incident_report]
       
@@ -101,15 +141,16 @@ class IncidentReportsController < ApplicationController
         if @incident_report.save
           format.html { redirect_to(@incident_report, :notice => 'Incident report was successfully created.') }
           format.xml  { render :xml => @incident_report, :status => :created, :location => @incident_report }
-          format.iphone {render :layout => 'mobile_application'}
+          #format.iphone {render :layout => 'mobile_application'}
+          format.iphone {redirect_to(@incident_report)}
         else
           format.html { render :action => "new_report" }
           format.xml  { render :xml => @incident_report.errors, :status => :unprocessable_entity }
-          format.iphone { render :layout => 'mobile_application'}
+          format.iphone { render :action => "new_report", :layout => 'mobile_application'}
         end
 			end   
   end
-  
+end
   
   
   
@@ -165,6 +206,13 @@ class IncidentReportsController < ApplicationController
     # get the report
     @incident_report = IncidentReport.find(params[:id])
     
+    # check authorization
+    if(Authorize.ra_authorize(current_staff) && current_staff != @incident_report.staff) || (Authorize.ra_authorize(current_staff) && current_staff == @incident_report.staff && @incident_report.submitted)
+        flash[:notice] = "Unauthorized Access"
+        redirect_to "/home/landingpage"
+        return
+    end
+    
     # destroy all reported infractions associated with it
     @incident_report.report_participant_relationships.each do |ri|
       ri.destroy
@@ -192,9 +240,6 @@ class IncidentReportsController < ApplicationController
   def new_report 	 
   	  	  
   	  logger.debug "inside IR new_report"
-    # incident_report in session will be nil if first visit to page
-
- #   if session[:incident_report] == nil
       @incident_report = IncidentReport.new                # new report
       @incident_report.staff_id = current_staff.id         # set submitter
       @annotation = Annotation.new                         # new annotation
@@ -202,18 +247,9 @@ class IncidentReportsController < ApplicationController
       #save everything to the session
       session[:incident_report] = @incident_report
       session[:annotation] = @annotation
-=begin      
-    else
 
-     # if incident report in session is not nil (not first visit)
-      # add students returned by search to report by creating fyi infractions
-      @incident_report = session[:incident_report]
-      #@incident_report.update_attributes_without_save(params)
-      @incident_report.add_default_report_student_relationships_for_participant_array(session[:students])
-      session[:students] = nil
-=end
-  	  	  # all above to on_initialize
-    
+   
+
     respond_to do |format|
       format.html # new_report.html.erb
       format.xml  { render :xml => @incident_report }
@@ -284,6 +320,11 @@ class IncidentReportsController < ApplicationController
   	@incident_report.add_default_report_student_relationships_for_participant_array([ @student ])
   	respond_to do |format|
    	   format.js
+   	   format.iphone {
+   	   render :update do |page|
+   	   	   page.replace_html("s-i-form", render( :partial => "student_infractions", :locals => { :ir => @incident_report }))
+   	   end
+   	   }
    	end 
   end
